# specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillHub",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/skills/{slug}/reviews": {
      "get": {
        "tags": [
          "social"
        ],
        "summary": "Get Reviews",
        "description": "Get paginated reviews sorted by helpful_count DESC.",
        "operationId": "get_reviews_api_v1_skills__slug__reviews_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "social"
        ],
        "summary": "Post Review",
        "description": "Create a review. 409 if already reviewed.",
        "operationId": "post_review_api_v1_skills__slug__reviews_post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/skills/{slug}/reviews/{review_id}": {
      "patch": {
        "tags": [
          "social"
        ],
        "summary": "Patch Review",
        "description": "Update a review. Owner only.",
        "operationId": "patch_review_api_v1_skills__slug__reviews__review_id__patch",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/submissions/{submission_id}/review": {
      "post": {
        "tags": [
          "submissions"
        ],
        "summary": "Review Submission Endpoint",
        "description": "Gate 3 human review. Platform team only.",
        "operationId": "review_submission_endpoint_api_v1_admin_submissions__submission_id__review_post",
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Submission Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Review Submission Endpoint Api V1 Admin Submissions  Submission Id  Review Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/access-requests/{request_id}/review": {
      "post": {
        "tags": [
          "submissions"
        ],
        "summary": "Review Access Request Endpoint",
        "description": "Review a division access request. Platform team only.",
        "operationId": "review_access_request_endpoint_api_v1_admin_access_requests__request_id__review_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessRequestReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Review Access Request Endpoint Api V1 Admin Access Requests  Request Id  Review Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/review-queue": {
      "get": {
        "tags": [
          "review-queue"
        ],
        "summary": "List Review Queue",
        "description": "List submissions awaiting human review.",
        "operationId": "list_review_queue_api_v1_admin_review_queue_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewQueueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/review-queue/{submission_id}/claim": {
      "post": {
        "tags": [
          "review-queue"
        ],
        "summary": "Claim",
        "description": "Claim a submission for review.",
        "operationId": "claim_api_v1_admin_review_queue__submission_id__claim_post",
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/review-queue/{submission_id}/decision": {
      "post": {
        "tags": [
          "review-queue"
        ],
        "summary": "Decide",
        "description": "Make a review decision on a submission.",
        "operationId": "decide_api_v1_admin_review_queue__submission_id__decision_post",
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Submission Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccessRequestReviewRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "pattern": "^(approved|denied)$",
            "title": "Decision"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "AccessRequestReviewRequest",
        "description": "Request body for reviewing an access request."
      },
...
      "ClaimResponse": {
        "properties": {
          "submission_id": {
            "type": "string",
            "title": "Submission Id"
          },
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id"
          },
          "claimed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Claimed At"
          }
        },
        "type": "object",
        "required": [
          "submission_id",
          "reviewer_id",
          "claimed_at"
        ],
        "title": "ClaimResponse",
        "description": "Response after claiming a submission for review."
      },
...
      "DecisionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "DecisionRequest",
        "description": "Request body for a review decision."
      },
      "DecisionResponse": {
        "properties": {
          "submission_id": {
            "type": "string",
            "title": "Submission Id"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id"
          },
          "reviewed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "submission_id",
          "decision",
          "reviewer_id",
          "reviewed_at"
        ],
        "title": "DecisionResponse",
        "description": "Response after making a review decision."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewCreateRequest": {
        "properties": {
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating"
          },
          "body": {
            "type": "string",
            "minLength": 1,
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "rating",
          "body"
        ],
        "title": "ReviewCreateRequest",
        "description": "Body for POST /skills/{slug}/reviews."
      },
      "ReviewDecisionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "pattern": "^(approved|changes_requested|rejected)$",
            "title": "Decision"
          },
          "notes": {
            "type": "string",
            "minLength": 1,
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "notes"
        ],
        "title": "ReviewDecisionRequest",
        "description": "Request body for Gate 3 human review."
      },
      "ReviewListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "per_page",
          "has_more"
        ],
        "title": "ReviewListResponse",
        "description": "Paginated list of reviews."
      },
      "ReviewQueueItem": {
        "properties": {
          "submission_id": {
            "type": "string",
            "title": "Submission Id"
          },
          "display_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Id"
          },
          "skill_name": {
            "type": "string",
            "title": "Skill Name"
          },
          "short_desc": {
            "type": "string",
            "title": "Short Desc"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "submitter_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitter Name"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "gate1_passed": {
            "type": "boolean",
            "title": "Gate1 Passed"
          },
          "gate2_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate2 Score"
          },
          "gate2_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate2 Summary"
          },
          "content_preview": {
            "type": "string",
            "title": "Content Preview"
          },
          "wait_time_hours": {
            "type": "number",
            "title": "Wait Time Hours"
          },
          "divisions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Divisions"
          }
        },
        "type": "object",
        "required": [
          "submission_id",
          "skill_name",
          "short_desc",
          "category",
          "gate1_passed",
          "content_preview",
          "wait_time_hours",
          "divisions"
        ],
        "title": "ReviewQueueItem",
        "description": "Single item in the review queue."
      },
      "ReviewQueueResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewQueueItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "per_page",
          "has_more"
        ],
        "title": "ReviewQueueResponse",
        "description": "Paginated review queue response."
      },
      "ReviewResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "skill_id": {
            "type": "string",
            "format": "uuid",
            "title": "Skill Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "helpful_count": {
            "type": "integer",
            "title": "Helpful Count"
          },
          "unhelpful_count": {
            "type": "integer",
            "title": "Unhelpful Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "skill_id",
          "user_id",
          "rating",
          "body",
          "helpful_count",
          "unhelpful_count",
          "created_at",
          "updated_at"
        ],
        "title": "ReviewResponse",
        "description": "Single review in response."
      },
      "ReviewUpdateRequest": {
        "properties": {
          "rating": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "body": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          }
        },
        "type": "object",
        "title": "ReviewUpdateRequest",
        "description": "Body for PATCH /skills/{slug}/reviews/{id}."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}